requirements.txt
mkdocs>=1.0
pandas>=1.1
tabulate>=0.8.7
PyYAML>=5.4.1